Use real-time scheduler for priority changes.
On unix-compatible OSes the real-time FIFO scheduler is now used for any
reasonable priority value. This is done to explicitly tell the user that
the non-zero values are unsupported in normal scheduling policies:

<https://man7.org/linux/man-pages/man7/sched.7.html>

>   For threads scheduled under one of the normal scheduling policies
>   (SCHED_OTHER, SCHED_IDLE, SCHED_BATCH), sched_priority is not
>   used in scheduling decisions (it must be specified as 0).

That means tha the priority value is only useful for real-time policies
and shouldn't be used with the normal ones. This is a breaking change in
the implementation and partially the API, so the version number is
bumped to 0.7.0.