Be more explicit about allowed priority values in POSIX.
This commit uses POSIX functions to get the maximum and minimum values
for different scheduling policies. Previously, the values of 0 and 99
were used as the range of allowed values, but according to the standard,
this isn't guaranteed to be always the same, hence such functions for
getting the values exist. To be more correct and tell the user what the
range is in case one specified an incorrect value, this commit makes use
of the mentioned API.